Move sidecar main() to direktiv main().
