+++ cmd/direktiv/main.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 	"github.com/direktiv/direktiv/cmd/api"
 	"github.com/direktiv/direktiv/cmd/secrets"
+	"github.com/direktiv/direktiv/cmd/sidecar"
 	"log"
 	"os"
 )
@@ -16,6 +17,8 @@ func main() {
 		api.RunApplication()
 	case "secrets":
 		secrets.RunApplication()
+	case "sidecar":
+		sidecar.RunApplication()
 	case "":
 		log.Fatalf("error: empty DIREKTIV_APP environment variable.\n")
 	default:
